Add NodeCollections
This adds collections and a collection manager to allow expressive use of groups to organize nodes.

It also adds the concept of a CollectionConstraint interface which is then implemented as MutuallyExclusivityConstraint.  This constraint is applied to both "tenant" and "partition" NodeCollectionTypes to ensure that a node can only be part of one tenant at a time and one partition at a time, but can be part of a partition and tenant collection.

The four built-in types ("partition", "tenant", "job", and "ad-hoc") are just a start to demonstrate how collection types work.