Added jwt parsing and logging
Both the server and client are now adapted to use jwts with a uid and sub field.  Both are used/stored when creating/updating a  collection and logged when working with nodes.

Added structured logging with logsrus to start thinking about wide events.  Including the request-id as well as the jwt subject of every request.

I'm not sure "sub" and "uid" are both necessary here.