feat: Add boot IPv4 and IPv6 address fields to ComputeNode
This commit adds two new fields, `BootIPv4Address` and `BootIPv6Address`, to the `ComputeNode` struct in the `nodes.go` file. These fields allow for specifying the boot IPv4 and IPv6 addresses for the node. This change enhances the functionality of the `ComputeNode` struct and enables more flexible network configuration for the nodes.